refactor: 영화 더 불러오기 함수 분리
